3.0.0 to production version

+++ Az.Avd/Public/Remove-AvdSessionHost.ps1
@@ -20,12 +20,12 @@ function Remove-AvdSessionHost {
         [parameter(Mandatory, ParameterSetName = 'Hostname')]
         [ValidateNotNullOrEmpty()]
         [string]$HostpoolName,
-    
+
         [parameter(Mandatory, ParameterSetName = 'All')]
         [parameter(Mandatory, ParameterSetName = 'Hostname')]
         [ValidateNotNullOrEmpty()]
         [string]$ResourceGroupName,
-    
+
         [parameter(Mandatory, ParameterSetName = 'Hostname')]
         [ValidateNotNullOrEmpty()]
         [string]$Name,
@@ -46,7 +46,6 @@ function Remove-AvdSessionHost {
     )
     Begin {
         Write-Verbose "Start removing sessionhosts"
-        AuthenticationCheck
         $sessionHostParameters = @{
             hostpoolName      = $HostpoolName
             resourceGroupName = $ResourceGroupName
@@ -92,20 +91,26 @@ function Remove-AvdSessionHost {
             }
             try {
                 if ($DeleteAssociated.IsPresent) {
-                    Write-Warning "Delete associated resources provided."
-                    Write-Verbose "Associated resources (disk & NIC) also will be removed"
-                    Write-Information "Looking for network resources" -InformationAction Continue
-                    $sh.vmResources.properties.networkprofile.networkInterfaces.id | ForEach-Object {
-                        Remove-Resource -resourceId $_ -apiVersion "2022-01-01"
-                    }
-                    Write-Information "Looking for OS disk" -InformationAction Continue
-                    Remove-Resource -resourceId $sh.vmResources.properties.storageProfile.osDisk.managedDisk.id -apiVersion "2022-03-02"
-                    Write-Information "Looking for data disks" -InformationAction Continue
-                    if ($sh.vmResources.properties.storageProfile.dataDisks.ManagedDisk) {
-                        Write-Verbose "Data disks found, removing them also"
-                        $sh.vmResources.properties.storageProfile.dataDisks.ManagedDisk.id | ForEach-Object {
-                            Remove-Resource -resourceId $_ -apiVersion "2022-03-02"
+                    # Deallocate first before deleting all resources
+                    if ($sh.vmresources.properties.instanceview.statuses.displayStatus -eq 'VM deallocated') {
+                        Write-Warning "Delete associated resources provided."
+                        Write-Verbose "Associated resources (disk & NIC) also will be removed"
+                        Write-Information "Looking for network resources" -InformationAction Continue
+                        $sh.vmResources.properties.networkprofile.networkInterfaces.id | ForEach-Object {
+                            Remove-Resource -resourceId $_ -apiVersion "2022-01-01"
                         }
+                        Write-Information "Looking for OS disk" -InformationAction Continue
+                        Remove-Resource -resourceId $sh.vmResources.properties.storageProfile.osDisk.managedDisk.id -apiVersion "2022-03-02"
+                        Write-Information "Looking for data disks" -InformationAction Continue
+                        if ($sh.vmResources.properties.storageProfile.dataDisks.ManagedDisk) {
+                            Write-Verbose "Data disks found, removing them also"
+                            $sh.vmResources.properties.storageProfile.dataDisks.ManagedDisk.id | ForEach-Object {
+                                Remove-Resource -resourceId $_ -apiVersion "2022-03-02"
+                            }
+                        }
+                    }
+                    else {
+                        Write-Warning "Skipped resources removal for session host $($sh.name), VM is not deallocated"
                     }
                 }
             }
